% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{CNVBurdenTest}
\alias{CNVBurdenTest}
\title{This function is to test for geneset burden of gene count.}
\usage{
CNVBurdenTest(
  cnv.matrix,
  geneset,
  label,
  covariates,
  correctGlobalBurden = T,
  correctTotal = T,
  standardizeCoefficient = T,
  permutation = T,
  nperm = 100,
  BiasedUrn = F
)
}
\arguments{
\item{cnv.matrix}{CNV table containing samples' outcome, gene count, covariates (optional) and cnv count (optional)}

\item{geneset}{a list of gene set, which each element contains a list of Entrez gene IDs}

\item{label}{variable name that is used as outcome}

\item{covariates}{list of covariates to be used in the model}

\item{correctGlobalBurden}{logical value to indicate whether the burden will be corrected for gene count or not: default (T)}

\item{correctTotal}{logical value to indicate whether the global burden should be corrected using total count or count outside gene set: default (T)}

\item{standardizeCoefficient}{logical value to indicate whether coefficient will be standardized or not: default (T)}

\item{permutation}{logical value to indicate whether permutation is required or not: default (T)}

\item{nperm}{numeric value to indicate number of iterations in permutation: default (100)}

\item{BiasedUrn}{logical value to indicate whether BaisedUrn will be used to permute label or not: default (F)}
}
\description{
This function is to test for geneset burden of gene count.
}
\keyword{GSBurden}
