% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{CNVGlobalTest}
\alias{CNVGlobalTest}
\title{This function is to test for global burden of gene count.}
\usage{
CNVGlobalTest(cnv.matrix, label, covariates, correctCNVCount = T,
  standardizeCoefficient = T)
}
\arguments{
\item{cnv.matrix}{CNV table containing samples' outcome, gene count, covariates (optional) and cnv count (optional)}

\item{label}{variable name that is used as outcome}

\item{covariates}{list of covariates to be used in the model}

\item{correctCNVCount}{logical value to indicate whether the burden will be corrected for CNV count or not}

\item{standardizeCoefficient}{logical value to indicate whether coefficient will be standardized or not}
}
\description{
This function is to test for global burden of gene count.
}
\keyword{GSBurden}
