% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{CNVLociTest}
\alias{CNVLociTest}
\title{This function is for loci test.}
\usage{
CNVLociTest(
  cnv.table,
  cnv.matrix,
  annotation.table,
  label,
  covariates,
  geneset = list(),
  permutation = T,
  nperm = 100,
  nsubject = 3,
  BiasedUrn = F
)
}
\arguments{
\item{cnv.table}{CNV table}

\item{cnv.matrix}{CNV matrix containing samples' outcome, gene count, covariates (optional) and cnv count (optional)}

\item{annotation.table}{gene annotation table}

\item{label}{variable name that is used as outcome}

\item{covariates}{list of covariates to be used in the model}

\item{geneset}{a list of gene set, which each element contains a list of Entrez gene IDs. If geneset parameter is specified, only genes in geneset will be tested}

\item{permutation}{logical, doing permutation FDR or not: default (T)}

\item{nperm}{number of permutation to be done: default (100)}

\item{nsubject}{minimum number of subjects with CNVs impacting the loci, required for a locus to be tested}

\item{BiasedUrn}{logical value to indicate whether BaisedUrn will be used to permute label or not: default (F)}
}
\description{
This function is for loci test.
}
\keyword{GSBurden}
