% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{CNVSet}
\alias{CNVSet}
\title{This function to get CNV in the format used by GSBurden.}
\usage{
CNVSet(sample, chr, start, end, type = "-")
}
\arguments{
\item{sample}{a vector of sample IDs}

\item{chr}{a vector of chromosome name of CNVs}

\item{start}{a vector of start locations of CNVs}

\item{end}{a vector of end locations of CNVs}

\item{type}{a vector of type of CNVs i.e. c("Loss", "Gain", "Loss", "Loss")}
}
\description{
This function to get CNV in the format used by GSBurden.
}
\keyword{Entrez}
\keyword{id}
