% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{GeneAnnotation}
\alias{GeneAnnotation}
\title{This function to get coordinate of gene for annotating CNVs/SNVs in the format used by GSBurden.}
\usage{
GeneAnnotation(enzid, chr, start, end, gsymbol = "-")
}
\arguments{
\item{enzid}{a vector of Entrez gene IDs}

\item{chr}{a vector of chromosome name of CNVs}

\item{start}{a vector of start locations of CNVs}

\item{end}{a vector of end locations of CNVs}
}
\description{
This function to get coordinate of gene for annotating CNVs/SNVs in the format used by GSBurden.
}
\keyword{Entrez}
\keyword{id}
