% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{getCNVDupGSMatrix}
\alias{getCNVDupGSMatrix}
\title{This function is to get a matrix of gene set count by sample but separate duplication by disruption.}
\usage{
getCNVDupGSMatrix(cnv.table, annotation.table, appris.cds,
  appris.nodisrupt, geneset)
}
\arguments{
\item{cnv.table}{CNV table}

\item{annotation.table}{gene annotation table}

\item{appris.cds}{appris principal isoform cds. Each row represent an exon but only coding sequence.}

\item{appris.nodisrupt}{appris principal isoform promoter and stop codon.}

\item{geneset}{gene set object}
}
\description{
This function is to get a matrix of gene set count by sample but separate duplication by disruption.
}
\keyword{GSBurden}
\keyword{Separate}
\keyword{by}
\keyword{disruption}
\keyword{duplication}
