% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{getCNVGSMatrix}
\alias{getCNVGSMatrix}
\title{This function is to get a matrix of gene set count by sample.}
\usage{
getCNVGSMatrix(cnv.table, annotation.table, geneset)
}
\arguments{
\item{cnv.table}{CNV table}

\item{annotation.table}{gene annotation table}

\item{geneset}{gene set object}
}
\description{
This function is to get a matrix of gene set count by sample.
}
\keyword{GSBurden}
