% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{getGenesetList}
\alias{getGenesetList}
\title{This function is for internal use by other functions in the package.}
\usage{
getGenesetList(enzid, geneset)
}
\arguments{
\item{enzid}{Entrez id}

\item{geneset}{geneset}
}
\description{
This function is for internal use by other functions in the package.
}
\keyword{GSBurden}
