% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{getSNVGSMatrix}
\alias{getSNVGSMatrix}
\title{This function is to get a matrix of gene set count by sample (variant count).}
\usage{
getSNVGSMatrix(snv.table, geneset)
}
\arguments{
\item{snv.table}{SNV table with sample(sample id), varid (unique variant id), entrez_id(entrez id) and type (type of snv)}

\item{geneset}{gene set object}
}
\description{
This function is to get a matrix of gene set count by sample (variant count).
}
\keyword{GSBurden}
