% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{readGeneset}
\alias{readGeneset}
\title{readGeneset Function}
\usage{
readGeneset(pathfile, sep = "\\t", header = T)
}
\arguments{
\item{pathfile}{path to gene set file. Either 1) A text file of two-column table; 1st column is Entrez gene ids and 2nd column is Gene set name. or 2) R object file, contain list variable. Each element is a vector contains Entrez gene ids. Gene set name is required as names of list elements.}
}
\description{
This function to read gene set information to perform gene set burden test for CNVs and SNVs.
}
\examples{
readGeneset('test.geneset.RData')
}
\keyword{analysis}
\keyword{entrez}
\keyword{gene}
\keyword{ids,}
\keyword{set}
\keyword{set,}
