% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{SNVBurdenTest}
\alias{SNVBurdenTest}
\title{This function is to test for geneset burden of snv count.}
\usage{
SNVBurdenTest(
  snv.matrix,
  geneset,
  label,
  covariates,
  correctGlobalBurden = T,
  standardizeCoefficient = T,
  permutation = T,
  nperm = 100,
  BiasedUrn = F
)
}
\arguments{
\item{snv.matrix}{SNV table containing samples' outcome, gene count, snv count for each gene set, covariates (optional) and total snv count}

\item{geneset}{a list of gene set, which each element contains a list of Entrez gene IDs}

\item{label}{variable name that is used as outcome}

\item{covariates}{list of covariates to be used in the model}

\item{correctGlobalBurden}{logical value to indicate whether the burden will be corrected for SNV count or not: default (T)}

\item{standardizeCoefficient}{logical value to indicate whether coefficient will be standardized or not: default (T)}

\item{permutation}{logical value to indicate whether permutation is required or not: default (T)}

\item{nperm}{numeric value to indicate number of iterations in permutation: default (100)}

\item{BiasedUrn}{logical value to indicate whether BaisedUrn will be used to permute label or not: default (F)}
}
\description{
This function is to test for geneset burden of snv count.
}
\keyword{GSBurden}
