% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{SNVGlobalTest}
\alias{SNVGlobalTest}
\title{This function is to test for global burden of gene count.}
\usage{
SNVGlobalTest(
  snv.matrix,
  label,
  covariates,
  correctSNVCount = F,
  standardizeCoefficient = T
)
}
\arguments{
\item{snv.matrix}{SNV table containing samples' outcome, gene count, snv count for each gene set, covariates (optional) and total snv count}

\item{label}{variable name that is used as outcome}

\item{covariates}{list of covariates to be used in the model}

\item{correctSNVCount}{logical value to indicate whether the burden will be corrected for SNV count or not: default (F)}

\item{standardizeCoefficient}{logical value to indicate whether coefficient will be standardized or not: default (T)}
}
\description{
This function is to test for global burden of gene count.
}
\keyword{GSBurden}
