% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{getCNVDupGSMatrix}
\alias{getCNVDupGSMatrix}
\title{This function is to get a matrix of gene set count by sample but separate duplication by disruption.}
\usage{
getCNVDupGSMatrix(cnv.table, annotation.table, appris.transcript, geneset)
}
\arguments{
\item{cnv.table}{CNV table}

\item{annotation.table}{gene annotation table}

\item{appris.transcript}{appris principal isoform transcript. Each row represent a full transcript.}

\item{geneset}{gene set object}
}
\description{
This function is to get a matrix of gene set count by sample but separate duplication by disruption.
}
\keyword{GSBurden}
\keyword{Separate}
\keyword{by}
\keyword{disruption}
\keyword{duplication}
