% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSBurden.function.R
\name{mergeLoci}
\alias{mergeLoci}
\title{This function is for internal use by other functions in the package.}
\usage{
mergeLoci(test.table, pvalue.column)
}
\description{
This function is for internal use by other functions in the package.
}
